react router dom installed so now there are different paths, had to move sign out functionality to its own file and change some of the values being passed through the children
